pyproject.toml
[tool.poetry]
name = "shedule-bot"
version = "0.1.0"
description = ""
authors = ["1220_school"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
telegram = "^0.0.1"
nest-asyncio = "^1.6.0"
aiosqlite = "^0.20.0"
python-telegram-bot = "^21.10"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
